+++ ui/src/main/kotlin/com/infinum/collar/ui/domain/entities/EntityRepository.kt
@@ -6,12 +6,14 @@ import com.infinum.collar.ui.domain.Repositories
 import com.infinum.collar.ui.domain.entities.models.EntityParameters
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
+import me.tatarka.inject.annotations.Inject
 
+@Inject
 internal class EntityRepository(
     private val dao: EntitiesDao
 ) : Repositories.Entity {
 
-    override suspend fun save(input: EntityParameters) =
+    override suspend fun save(input: EntityParameters): Long =
         input.entity?.let { dao.save(it) }
             ?: error("Cannot save null entity")
 
@@ -29,6 +31,9 @@ internal class EntityRepository(
                 else -> dao.load()
             }
         }
+
+    override suspend fun loadById(input: EntityParameters): List<CollarEntity> =
+        listOfNotNull(input.entityId?.let { dao.load(it) })
 
     override suspend fun clear() =
         dao.delete()