+++ ui/src/main/kotlin/com/infinum/collar/ui/domain/shared/base/BaseRepository.kt
@@ -4,7 +4,11 @@ import kotlinx.coroutines.flow.Flow
 
 internal interface BaseRepository<InputModel : BaseParameters, OutputModel> {
 
-    suspend fun save(input: InputModel): Unit = throw NotImplementedError()
+    suspend fun save(input: InputModel): Long = throw NotImplementedError()
 
     suspend fun load(input: InputModel): Flow<OutputModel> = throw NotImplementedError()
+
+    suspend fun loadById(input: InputModel): OutputModel = throw NotImplementedError()
+
+    suspend fun clear(): Unit = throw NotImplementedError()
 }