+++ ui/src/main/kotlin/com/infinum/collar/ui/presentation/CollarEvent.kt
@@ -1,6 +1,12 @@
 package com.infinum.collar.ui.presentation
 
+import com.infinum.collar.ui.data.models.local.CollarEntity
+
 internal sealed class CollarEvent {
+
+    data class ShowEntity(
+        val entity: CollarEntity
+    ) : CollarEvent()
 
     data class Filters(
         val screens: Boolean,